CUSTOMER AND BUSINESS DATA GENERATOR
This is a revised project that I initially created in 2018. Initially the project was to generate names and other personal data. Recently I also added parts, orders and invoices (a table that combines parts to customer orders). I use this data base to generate data that I use to practice SQL and other data analysis tools.